Handle "processing" status by retrying as the RFC suggests.
Previously we would give up with "unhandled status" and not get a
certificate.

While here limit the amount of retries we are willing to perform in
the "pending" and "processing" state.

Let's encrypt seems to enter the "processing" state farely often in
the staging environment.

from florian@; problem reported & OK sthen

this is errata/7.7/005_acme.patch.sig